+++ contracts/interfaces/IGasPriceOracle.sol
@@ -3,12 +3,12 @@ pragma solidity 0.8.7;
 
 interface IGasPriceOracle {
     function relativeGasPrice(
-        uint256 dstChainSlug
+        uint32 dstChainSlug
     ) external view returns (uint256);
 
     function sourceGasPrice() external view returns (uint256);
 
     function getGasPrices(
-        uint256 dstChainSlug_
+        uint32 dstChainSlug_
     ) external view returns (uint256, uint256);
 }
