+++ contracts/switchboard/default-switchboards/OptimisticSwitchboard.sol
@@ -26,7 +26,7 @@ contract OptimisticSwitchboard is SwitchboardBase {
     function allowPacket(
         bytes32,
         bytes32,
-        uint256 srcChainSlug_,
+        uint32 srcChainSlug_,
         uint256 proposeTime_
     ) external view override returns (bool) {
         if (tripGlobalFuse || tripSinglePath[srcChainSlug_]) return false;
