+++ contracts/switchboard/native/OptimismSwitchboard.sol
@@ -67,7 +67,7 @@ contract OptimismSwitchboard is NativeSwitchboardBase, INativeReceiver {
     }
 
     function initateNativeConfirmation(bytes32 packetId_) external {
-        uint256 capacitorPacketCount = uint256(uint64(uint256(packetId_)));
+        uint64 capacitorPacketCount = uint64(uint256(packetId_));
         bytes32 root = capacitor__.getRootByCount(capacitorPacketCount);
         bytes memory data = abi.encodeWithSelector(
             INativeReceiver.receivePacket.selector,
@@ -98,7 +98,7 @@ contract OptimismSwitchboard is NativeSwitchboardBase, INativeReceiver {
     function allowPacket(
         bytes32 root_,
         bytes32 packetId_,
-        uint256,
+        uint32,
         uint256
     ) external view override returns (bool) {
         if (tripGlobalFuse) return false;
